t5

+++ 0x07-pointers_arrays_strings/8-print_diagsums.c
@@ -10,19 +10,19 @@
  
 void print_diagsums(int *a, int size)
 {
-        int i, sum;
+        long i, sum;
 
         for (i = 0; i < size; i++)
         {
                 sum += a[i * size + i];
         }
-        printf("%d, ", sum);
+        printf("%ld, ", sum);
         sum = 0;
         for (i = 0; i < size; i++)
         {
                 sum += a[i * size + size - i - 1];               
         }
-	printf("%d\n", sum);
+	printf("%ld\n", sum);
 }
